% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weo.R
\name{read_weo}
\alias{read_weo}
\title{Read data from the IMF World Economic Outlook}
\usage{
read_weo(month_year, path = tempdir())
}
\arguments{
\item{month_year}{Character string such as "Oct 2022" or "October 2022"
or "April 2021".}

\item{path}{Path to directory where downloaded file should be stored}
}
\description{
The IMF releases World Economic Outlook reports typically twice a year,
in April and October. These reports contain a range of data and forecasts of
key economic variables. They provide this data (including the forecasts) in
downloadable tables. \code{read_weo()} downloads, imports, and tidies these
tables.
}
\examples{
\dontrun{
read_weo("Oct 2022")

read_weo("April 2019")
}
}
